refactor(api): migrate route for /api/admin/users/{id}/organizations
