+++ api/lib/application/users/index.js
@@ -8,34 +8,6 @@ const register = async function (server) {
   const adminRoutes = [
     {
       method: 'GET',
-      path: '/api/admin/users/{id}/organizations',
-      config: {
-        pre: [
-          {
-            method: (request, h) =>
-              securityPreHandlers.hasAtLeastOneAccessOf([
-                securityPreHandlers.checkAdminMemberHasRoleSuperAdmin,
-                securityPreHandlers.checkAdminMemberHasRoleCertif,
-                securityPreHandlers.checkAdminMemberHasRoleSupport,
-                securityPreHandlers.checkAdminMemberHasRoleMetier,
-              ])(request, h),
-          },
-        ],
-        validate: {
-          params: Joi.object({
-            id: identifiersType.userId,
-          }),
-        },
-        handler: userController.findUserOrganizationsForAdmin,
-        notes: [
-          "- **Cette route est restreinte aux utilisateurs authentifiés ayant les droits d'accès**\n" +
-            '- Elle permet à un administrateur de lister les organisations auxquelles appartient l´utilisateur',
-        ],
-        tags: ['api', 'admin', 'user', 'organizations'],
-      },
-    },
-    {
-      method: 'GET',
       path: '/api/admin/users/{id}/certification-center-memberships',
       config: {
         pre: [
