+++ api/tests/team/acceptance/application/membership/membership.admin.route.test.js
@@ -15,6 +15,40 @@ describe('Acceptance | Team | Admin | Route | membership', function () {
     server = await createServer();
   });
 
+  describe('GET /api/admin/users/{id}/organizations', function () {
+    it('returns 200 HTTP status code', async function () {
+      // given
+      const server = await createServer();
+      const userId = databaseBuilder.factory.buildUser().id;
+
+      const organization1 = databaseBuilder.factory.buildOrganization({
+        name: 'Organization 1',
+      });
+
+      databaseBuilder.factory.buildMembership({
+        organizationId: organization1.id,
+        userId,
+        organizationRole: 'MEMBER',
+      });
+
+      const admin = databaseBuilder.factory.buildUser.withRole();
+
+      await databaseBuilder.commit();
+
+      const options = {
+        method: 'GET',
+        url: `/api/admin/users/${userId}/organizations`,
+        headers: generateAuthenticatedUserRequestHeaders({ userId: admin.id }),
+      };
+
+      // when
+      const response = await server.inject(options);
+
+      // then
+      expect(response.statusCode).to.equal(200);
+    });
+  });
+
   describe('POST /api/admin/memberships', function () {
     let options;
     let userId;
